local script separated from server stuff
