+++ local/main.py
@@ -1,0 +1,4 @@
+from add_card import add_card
+
+if __name__ == "__main__":
+    add_card()
