+++ local/scripts/__init__.py
@@ -1,0 +1,3 @@
+from scripts.request_wrapper import send_response
+from scripts.add_card_scripts import add_append_query, add_parse_list, add_select_card
+
