+++ local/scripts/add_card_scripts/parse_response.py
