--- scripts/__init__.py
@@ -1,8 +1,0 @@
-# import scripts.add_card as add_card
-# import scripts.query_price as query_price
-from scripts.util import  util_check_promo,  util_sanitize_string,  util_create_directory,  util_send_response, util_connect_to_db
-from scripts.add_card_scripts import add_parse_list, add_select_card, add_append_query
-from scripts.query_price_scripts import query_add_data
-
-from scripts.add_card import add_card
-from scripts.query_price import query_price
