--- scripts/util/__init__.py
@@ -1,7 +1,0 @@
-# Do I really need to care about always importing all these utility scripts?
-
-from scripts.util.sanitize_string import sanitize_string as util_sanitize_string
-from scripts.util.create_directory import create_directory as util_create_directory
-from scripts.util.request_wrapper import send_response as util_send_response
-from scripts.util.check_set_promo import check_set_promo as util_check_promo
-from scripts.util.db_connect import _connect_to_db as util_connect_to_db
